refactor: move four CLI option/help leaves into frontends (ADR-061 Phase 4)
Phase 4's recorded blocker was that the option-declaration cluster "is not
leaf-shaped ... these ~3,800 combined lines import each other", needing "its
own pass that first splits the cluster's internal dependency graph".

Re-measured before scoping against it, as this ADR's own Phase 2 lesson says
to -- the third time here a recorded blocker has described a tree that had
since moved. An AST scan of the six modules found a star, not a tangle: five
of them (cli_params, cli_profiles, cli_options_contract, cli_contract_options,
cli_help) have zero intra-cluster imports; only the cli_options.py hub imports
them. There was no internal dependency graph left to split.

Four of the five also have zero first-party imports of any kind -- exactly the
criterion secondary_output.py was moved on -- so they moved, 1,249 lines with
six call sites:

  cli_profiles.py         -> frontends/cli/options/profiles.py
  cli_contract_options.py -> frontends/cli/options/contract.py
  cli_options_contract.py -> frontends/cli/options/inventory.py
  cli_help.py             -> frontends/cli/help.py

The two renames are deliberate rather than incidental. cli_options_contract
(the cli-contract gate's flag inventory and budget ledger) and
cli_contract_options (ADR-049's contract-evaluation options) are unrelated
things whose names differed only by word order -- precisely the "physical
ownership is ambiguous" problem in this ADR's own Context.

cli_params.py (452 lines) stays flat, and the reason is the rule worth
remembering: frontends/ is a *migrated* package, so unclassified-import
applies to anything physically inside it. cli_params imports four unclassified
flat modules (policies, policy_file, suppression, buildsource.scan_levels); it
moves once those have owners. Recorded in frontends/AGENTS.md, whose own
"cluster modules import each other" claim is corrected.

Verified along the way, rather than reported as a bug: find_packages() does
not see abicheck.frontends (no __init__.py in frontends/ or frontends/cli/),
which looked like the Phase 4 package being missing from the wheel. Built a
wheel to check -- the files are there. [tool.setuptools.packages.find]
defaults to namespaces = true, unlike the legacy find_packages() I probed
with. No packaging bug; noted so the next reader does not re-raise it.

One self-inflicted regression, caught by the suite and worth recording since
it is the second time: `ruff check --fix` on scripts/ stripped MODULE_SCOPE
from check_mutation_score.py's import block as "unused" -- it is a deliberate
re-export that tests/test_mutation_score_gate.py reads as gate.MODULE_SCOPE.
Restored with an `as` re-export so a future --fix cannot strip it again. Do
not run ruff --fix broadly on this repo; it removes intentional re-exports.

Two machine-checked registries carried the old paths and are updated
(docs/_meta/topics.yaml fact_sources, docs/contribute/usecase-registry.yaml
modules), plus scripts/skill_eval_surface.py's hard-coded path. Narrative
mentions in plan/ADR history are left as-is, per this repo's convention.
