+++ abicheck/cli_options.py
@@ -1836,7 +1836,7 @@ build_source_compare_options = evidence_options
 #: cap. Re-exported here so every existing caller — the ``cli-contract``
 #: gate's own tests and ``tests/test_config_rebalance.py`` — keeps its
 #: import path, the same pattern used for ``cli_profiles.py`` below.
-from .cli_options_contract import (  # noqa: E402
+from .frontends.cli.options.inventory import (  # noqa: E402
     COMPARE_FLAG_BUDGET as COMPARE_FLAG_BUDGET,
     COMPARE_FLAG_BUDGET_BASE as COMPARE_FLAG_BUDGET_BASE,
     COMPARE_FLAG_BUDGET_RAISES as COMPARE_FLAG_BUDGET_RAISES,
@@ -1865,7 +1865,7 @@ from .cli_options_contract import (  # noqa: E402
 # Spelled ``X as X`` (an explicit re-export) rather than declared in an
 # ``__all__``: this module has never had one, and adding a three-name list
 # would quietly narrow what ``import *`` gives every other consumer.
-from .cli_profiles import (  # noqa: E402
+from .frontends.cli.options.profiles import (  # noqa: E402
     COMPARE_PROFILES as COMPARE_PROFILES,
     RUN_PROFILE_META_KEY as RUN_PROFILE_META_KEY,
     profile_option as profile_option,
@@ -1971,7 +1971,7 @@ def apply_compare_profile(ctx: object, kwargs: dict[str, object]) -> None:
 #: before it. Re-exported here (``X as X``, so the re-export is explicit to
 #: mypy) because it is a shared decorator callers already reach through this
 #: module, and a leaf holding option definitions never imports back.
-from .cli_contract_options import (  # noqa: E402
+from .frontends.cli.options.contract import (  # noqa: E402
     contract_options as contract_options,
     pack_option as pack_option,
 )
