+++ abicheck/frontends/cli/options/inventory.py
