+++ docs/contribute/usecase-registry.yaml
@@ -396,7 +396,7 @@ use_cases:
     gap: G21
     plan: docs/contribute/plans/g21-oneshot-deep-compare.md
     evidence:
-      modules: [abicheck/cli.py, abicheck/cli_help.py, abicheck/cli_dump_helpers.py]
+      modules: [abicheck/cli.py, abicheck/frontends/cli/help.py, abicheck/cli_dump_helpers.py]
       tests: [tests/test_depth_vocabulary.py, tests/test_cov95_cli.py]
     next_steps: >
       Superseded: the standalone `deep-compare` orchestrator was removed and
